% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_lexicon}
\alias{check_lexicon}
\title{Ensure specificity for certain functions}
\usage{
check_lexicon(lexicon)
}
\arguments{
\item{lexicon}{a lexicon to be evaluated}
}
\value{
error if lexicon is not properly formatted
}
\description{
Ensure specificity for certain functions
}
