% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_lexicon_name}
\alias{check_lexicon_name}
\title{Checks lexicon name passed to \code{lexcion}}
\usage{
check_lexicon_name(lexicon_name)
}
\arguments{
\item{lexicon_name}{the lexicon name passed to \code{lexcion}}
}
\value{
error and message if lexicon not in available lexicons
}
\description{
Checks lexicon name passed to \code{lexcion}
}
