% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_named}
\alias{check_named}
\title{Check to see if needed parameters are missing}
\usage{
check_named(list)
}
\arguments{
\item{list}{a list to check}
}
\value{
error if list is not named
}
\description{
Check to see if needed parameters are missing
}
