% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_null}
\alias{check_null}
\title{Check to see if needed parameters are missing}
\usage{
check_null(...)
}
\arguments{
\item{...}{named items to check}
}
\value{
error if null
}
\description{
Check to see if needed parameters are missing
}
