% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_specific}
\alias{check_specific}
\title{Ensure specificity for certain functions}
\usage{
check_specific(v, name = NULL)
}
\arguments{
\item{v}{named items to check}

\item{name}{name of the argument being checked}
}
\value{
error if list is not named
}
\description{
Ensure specificity for certain functions
}
