% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giftwrap.R
\name{create_giftwrap}
\alias{create_giftwrap}
\title{giftwrap factory function}
\usage{
create_giftwrap(command, env, base_remove = NULL)
}
\arguments{
\item{command}{the shell command to be converted into an R function}

\item{env}{the environment in which the function should be created}

\item{base_remove}{remove the base from the function name by adding the base name here}
}
\value{
A function exported to the specified environment
}
\description{
giftwrap factory function
}
