% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giftwrap.R
\name{format_arg}
\alias{format_arg}
\title{Create an argument with shell formatting}
\usage{
format_arg(x)
}
\arguments{
\item{x}{the argument to be passed to the shell command. Both named arguments and unnamed arguments work. A list is expected.}
}
\value{
a shell-formatted argument
}
\description{
Create an argument with shell formatting
}
