% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giftwrap.R
\name{format_args}
\alias{format_args}
\title{format multiple arguments}
\usage{
format_args(...)
}
\arguments{
\item{...}{named and unnamed arguments to be formatted}
}
\value{
a vector of properly formatted arguments
}
\description{
format multiple arguments
}
