% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giftwrap.R
\name{giftwrap_command}
\alias{giftwrap_command}
\title{giftwrap a base command and its arguments from R to shell}
\usage{
giftwrap_command(base_command, ...)
}
\arguments{
\item{base_command}{a shell command}

\item{...}{named and unnamed arguments to be giftwrapped}
}
\value{
a shell-formatted command with shell-formatted arguments
}
\description{
giftwrap a base command and its arguments from R to shell
}
