% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.R
\name{lexicon}
\alias{lexicon}
\title{For fetching a giftwrap lexicon}
\usage{
lexicon(lexicon)
}
\arguments{
\item{lexicon}{The name of a giftwrap lexicon}
}
\value{
a dataframe of lexicon
}
\description{
For fetching a giftwrap lexicon
}
