% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lexicon_az}
\alias{lexicon_az}
\title{A giftwrap-formatted lexicon of the majority of Microsoft Azure commands}
\format{
A dataset containing all base, commands, and subcommands from Microsoft Azure (az)
\describe{
\item{base}{the base, always az}
\item{command}{the command/service, such as sql}
\item{subcommand}{the subcommand for the command/service, such as create}
\item{giftwrap_command}{the base, command, and subcommand together}
...
}
}
\source{
made by scraping Microsoft Azure command line help docs
}
\usage{
lexicon_az
}
\description{
A dataset containing all base, commands, and subcommands from Microsoft Azure (az)
}
\keyword{datasets}
