% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lexicon_docker}
\alias{lexicon_docker}
\title{A giftwrap-formatted lexicon of all docker commands}
\format{
A dataset containing all base, commands, and subcommands from docker
\describe{
\item{base}{the base, always docker}
\item{command}{the command/service, such as ps or image}
\item{subcommand}{the subcommand for the command/service, such as ls}
\item{giftwrap_command}{the base, command, and subcommand together}
...
}
}
\source{
made by scraping docker help docs
}
\usage{
lexicon_docker
}
\description{
A dataset containing all base, commands, and subcommands from docker
}
\keyword{datasets}
