% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lexicon_gcloud}
\alias{lexicon_gcloud}
\title{A giftwrap-formatted lexicon of all Google Cloud Platform gcloud commands}
\format{
A dataset containing all base, commands, and subcommands from gcloud
\describe{
\item{base}{the base, always gcloud}
\item{command}{the command/service, such as compute}
\item{subcommand}{the subcommand for the command/service, such as ssh}
\item{giftwrap_command}{the base, command, and subcommand together}
...
}
}
\source{
made by scraping gcloud help docs
}
\usage{
lexicon_gcloud
}
\description{
A dataset containing all base, commands, and subcommands from gcloud
}
\keyword{datasets}
