% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lexicon_gh}
\alias{lexicon_gh}
\title{A giftwrap-formatted lexicon of all github CLI commands}
\format{
A dataset containing all base, commands, and subcommands from github CLI (gh)
\describe{
\item{base}{the base, always gh}
\item{command}{the command/service, such as pipelines}
\item{subcommand}{the subcommand for the command/service, such as add}
\item{giftwrap_command}{the base, command, and subcommand together}
...
}
}
\source{
made by scraping heroku help docs
}
\usage{
lexicon_gh
}
\description{
A dataset containing all base, commands, and subcommands from github CLI (gh)
}
\keyword{datasets}
