% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lexicon_git}
\alias{lexicon_git}
\title{A giftwrap-formatted lexicon of all git commands}
\format{
A dataset containing all base, commands, and subcommands from git
\describe{
\item{base}{the base, always git}
\item{command}{the command, such as init, pull, or status}
\item{subcommand}{NA for git lexicon}
\item{giftwrap_command}{the base, command, and subcommand together}
...
}
}
\source{
made by scraping git help docs
}
\usage{
lexicon_git
}
\description{
A dataset containing all base, commands, and subcommands from git
}
\keyword{datasets}
