% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lexicon_heroku}
\alias{lexicon_heroku}
\title{A giftwrap-formatted lexicon of all heroku commands}
\format{
A dataset containing all base, commands, and subcommands from heroku
\describe{
\item{base}{the base, always heroku}
\item{command}{the command/service, such as pipelines}
\item{subcommand}{the subcommand for the command/service, such as add}
\item{giftwrap_command}{the base, command, and subcommand together}
...
}
}
\source{
made by scraping heroku help docs
}
\usage{
lexicon_heroku
}
\description{
A dataset containing all base, commands, and subcommands from heroku
}
\keyword{datasets}
