% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lexicon_kubectl}
\alias{lexicon_kubectl}
\title{A giftwrap-formatted lexicon of all Kubernetes kubectl commands}
\format{
A dataset containing all base, commands, and subcommands from kubectl
\describe{
\item{base}{the base, always kubectl}
\item{command}{the command/service, such as ps or image}
\item{subcommand}{the subcommand for the command/service, such as ls}
\item{giftwrap_command}{the base, command, and subcommand together}
...
}
}
\source{
made by scraping kubectl help docs
}
\usage{
lexicon_kubectl
}
\description{
A dataset containing all base, commands, and subcommands from kubectl
}
\keyword{datasets}
