% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lexicon_sfdx_force}
\alias{lexicon_sfdx_force}
\title{A giftwrap-formatted lexicon of all Salesforce developer tools (sfdx force) commands}
\format{
A dataset containing all base, commands, and subcommands from Salesforce developer tools (sfdx force)
\describe{
\item{base}{the base, always 'sfdc force'}
\item{command}{the command/service, such as data}
\item{subcommand}{the subcommand for the command/service, such as bulk:upsert}
\item{giftwrap_command}{the base, command, and subcommand together, specially formatted for sfdx}
...
}
}
\source{
made by scraping docker help docs
}
\usage{
lexicon_sfdx_force
}
\description{
A dataset containing all base, commands, and subcommands from Salesforce developer tools (sfdx force)
}
\keyword{datasets}
