% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giftwrap.R
\name{load_lexicon}
\alias{load_lexicon}
\title{Load a functions from a lexicon into an environment}
\usage{
load_lexicon(lexicon, commands = NULL, subcommands = NULL, drop_base = F, env)
}
\arguments{
\item{lexicon}{a dataframe containing columns for base, command, subcommand, and giftwrap_command (the full command)}

\item{commands}{regex filtering for any commands in the lexicon}

\item{subcommands}{regex filtering for any subcommands in the lexicon}

\item{drop_base}{drop the base-level command from the caller function}

\item{env}{the environment into which the giftwrap functions should be exported}
}
\value{
Functions exported to the specified environment
}
\description{
Load a functions from a lexicon into an environment
}
