% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giftwrap.R
\name{run_command}
\alias{run_command}
\title{Run command}
\usage{
run_command(command, collect = FALSE)
}
\arguments{
\item{command}{a shell command}

\item{collect}{whether the output of the shell command should be collected in R}
}
\value{
messages from the running command, errors if failure
}
\description{
Run command
}
