% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giftwrap.R
\name{wrap_commands}
\alias{wrap_commands}
\title{To convert shell commands into giftwrapped functions}
\usage{
wrap_commands(
  ...,
  env = parent.frame(),
  base_remove = NULL,
  use_namespace = NULL
)
}
\arguments{
\item{...}{shell commands to be turned into giftwrapped functions}

\item{env}{the environment in which the function should be created}

\item{base_remove}{remove the base from the function name by adding the base name here}

\item{use_namespace}{a character string of a namespace to use for the resulting functions}
}
\value{
A function or functions exported to the specified environment
}
\description{
To convert shell commands into giftwrapped functions
}
