% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giftwrap.R
\name{wrap_lexicon}
\alias{wrap_lexicon}
\title{Load a functions from a lexicon into an environment}
\usage{
wrap_lexicon(
  lexicon,
  commands = NULL,
  subcommands = NULL,
  drop_base = FALSE,
  env = parent.frame(),
  use_namespace = NULL
)
}
\arguments{
\item{lexicon}{a dataframe containing columns for base, command, subcommand, and giftwrap_command (the full command)}

\item{commands}{regex filtering for any commands in the lexicon}

\item{subcommands}{regex filtering for any subcommands in the lexicon}

\item{drop_base}{drop the base-level command from the caller function}

\item{env}{the environment into which the giftwrap functions should be exported}

\item{use_namespace}{a character string of a namespace for giftwrap to create and export the lexicon functions into}
}
\value{
Functions exported to the specified environment
}
\description{
Load a functions from a lexicon into an environment
}
