% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giftwrap.R
\name{giftwrap}
\alias{giftwrap}
\title{The caller function that will live inside of factory function}
\usage{
giftwrap(command, ..., process_echo = TRUE)
}
\arguments{
\item{command}{a shell command}

\item{...}{named and unnamed arguments to be giftwrapped}

\item{process_echo}{set to FALSE to suppress terminal messages when command is run}
}
\value{
messages from the running command, errors if failure
}
\description{
The caller function that will live inside of factory function
}
